task2/requirements.txt
langchain-community==0.3.31
langchain-core==0.3.85
langchain-gigachat==0.3.12
langchain-text-splitters==0.3.11
langgraph==1.1.10
numexpr==2.14.1
pypdf==6.10.2
python-dotenv==1.1.0
sentence-transformers==5.4.1
langchain==0.3.29
faiss-cpu==1.13.2
aiogram==3.27.0
